+++ protocol/x/vault/keeper/msg_server.go
@@ -5,12 +5,12 @@ import (
 )
 
 type msgServer struct {
-	Keeper types.VaultKeeper
+	Keeper
 }
 
 // NewMsgServerImpl returns an implementation of the MsgServer interface
 // for the provided Keeper.
-func NewMsgServerImpl(keeper types.VaultKeeper) types.MsgServer {
+func NewMsgServerImpl(keeper Keeper) types.MsgServer {
 	return &msgServer{Keeper: keeper}
 }
